Fill template mathswe-app for tsd.math.software
