+++ tsd.math.software---mvp/README.md
@@ -1,4 +1,22 @@
-# React + TypeScript + Vite
+<!-- Copyright (c) 2024 Tobias Briones. All rights reserved. -->
+<!-- SPDX-License-Identifier: CC-BY-4.0 -->
+<!-- This file is part of https://github.com/texsydo/texsydo---mvp -->
+
+# MVP: Tsd.Math.Software
+
+[![Project](https://mathswe-ops-services.tobiasbriones-dev.workers.dev/badge/project/texsydo)](https://tsd.math.software)
+&nbsp;
+[![GitHub Repository](https://img.shields.io/static/v1?label=GITHUB&message=REPOSITORY&labelColor=555&color=0277bd&style=for-the-badge&logo=GITHUB)](https://github.com/texsydo/texsydo---mvp/blob/main/tsd.math.software---mvp)
+
+[![GitHub Project License](https://img.shields.io/github/license/texsydo/texsydo---mvp.svg?style=flat-square)](https://github.com/texsydo/texsydo---mvp/blob/main/LICENSE)
+
+[![GitHub Release](https://mathswe-ops-services.tobiasbriones-dev.workers.dev/badge/version/github/texsydo/texsydo---mvp)](https://github.com/texsydo/texsydo---mvp/releases/latest)
+
+⚙ Production
+[![Netlify Status](https://api.netlify.com/api/v1/badges/435a4b07-f3ad-4490-a5b7-1d30f0f24834/deploy-status)](https://app.netlify.com/sites/texsydo/deploys)
+
+MVP of Texsydo (Textual System Documenting) web application that provides
+software for mathematical text and art.
 
 ## Getting Started
 
@@ -6,8 +24,24 @@ Install project dependencies via `npm install`.
 
 Run development mode via `npm run dev`.
 
-Run tests via `npm run test`.
-
 Build for production via `npm run build`.
 
-Run ESLint via `npx eslint .` or `npm run lint`.
+Run ESLint via `npx eslint .`.
+
+## Contact
+
+Tobias Briones: [GitHub](https://github.com/tobiasbriones)
+[LinkedIn](https://linkedin.com/in/tobiasbriones)
+
+## About
+
+**MVP: Tsd.Math.Software**
+
+Web application that provides software for mathematical text and art.
+
+Copyright © 2024 Tobias Briones. All rights reserved.
+
+### License
+
+This project is licensed under
+the [GNU Affero General Public License v3.0 or later License](../LICENSE).
