+++ tsd.math.software---mvp/index.html
@@ -1,16 +1,51 @@
+<!-- Copyright (c) 2024 Tobias Briones. All rights reserved. -->
+<!-- SPDX-License-Identifier: AGPL-3.0-or-later -->
+<!-- This file is part of https://github.com/texsydo/texsydo---mvp -->
+
 <!doctype html>
-<html lang="en">
+<html lang="en" dir="ltr">
     <head>
         <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <meta
+            name="title"
+            content="Texsydo"
+        />
+        <meta
+            name="description"
+            content="Texsydo (Textual System Documenting) provides mathematical text and art."
+        />
+        <meta
+            name="keywords"
+            content="Mathematical Software, Math, Applied, Art, Documentation, Article, Model, Rendering"
+        />
+        <meta
+            name="author"
+            content="Tobias Briones"
+        />
+        <meta
+            name="viewport"
+            content="width=device-width, initial-scale=1.0"
+        />
 
-        <title>Vite + React + TS</title>
+        <meta name="theme-color" content="#ffffff" />
+        <meta name="apple-mobile-web-app-capable" content="yes" />
 
-        <link href="/vite.svg" rel="icon" type="image/svg+xml" />
+        <title>Texsydo</title>
+
+        <link rel="icon" type="image/svg+xml" href="/texsydo.svg" />
+        <link
+            rel="apple-touch-icon" sizes="192x192"
+            href="/texsydo-192.png"
+        />
+        <link rel="preconnect" href="https://fonts.gstatic.com" />
+        <link
+            rel="stylesheet"
+            href="https://fonts.googleapis.com/css2?family=Poppins&display=swap"
+        />
     </head>
 
     <body>
         <div id="root"></div>
-        <script src="/src/main.tsx" type="module"></script>
+        <script src="/src/app/main.tsx" type="module"></script>
     </body>
 </html>
