+++ tsd.math.software---mvp/package.json
@@ -1,7 +1,7 @@
 {
-    "name": "react-ts-swc-vite",
+    "name": "tsd.math.software---mvp",
     "private": true,
-    "version": "0.0.0",
+    "version": "0.1.0-dev",
     "type": "module",
     "scripts": {
         "dev": "vite",
@@ -11,7 +11,12 @@
         "test": "vitest"
     },
     "dependencies": {
+        "@fortawesome/free-brands-svg-icons": "^6.6.0",
+        "@fortawesome/react-fontawesome": "^0.2.2",
+        "better-react-mathjax": "^2.0.3",
+        "bootstrap": "^5.3.3",
         "react": "^18.3.1",
+        "react-bootstrap": "^2.10.4",
         "react-dom": "^18.3.1"
     },
     "devDependencies": {
