--- tsd.math.software---mvp/src/App.tsx
@@ -1,39 +1,0 @@
-import viteLogo from "/vite.svg";
-import { useState } from "react";
-import reactLogo from "./assets/react.svg";
-import "./App.css";
-
-function App() {
-    const [ count, setCount ] = useState(0);
-
-    return (
-        <>
-            <div>
-                <a href="https://vitejs.dev" target="_blank" rel="noreferrer">
-                    <img src={ viteLogo } className="logo" alt="Vite logo" />
-                </a>
-                <a href="https://react.dev" target="_blank" rel="noreferrer">
-                    <img
-                        src={ reactLogo }
-                        className="logo react"
-                        alt="React logo"
-                    />
-                </a>
-            </div>
-            <h1>Vite + React</h1>
-            <div className="card">
-                <button onClick={ () => { setCount((count) => count + 1); } }>
-                    count is { count }
-                </button>
-                <p>
-                    Edit <code>src/App.tsx</code> and save to test HMR
-                </p>
-            </div>
-            <p className="read-the-docs">
-                Click on the Vite and React logos to learn more
-            </p>
-        </>
-    );
-}
-
-export default App;
