+++ tsd.math.software---mvp/src/app/App.css
@@ -1,0 +1,3 @@
+/* Copyright (c) 2024 Tobias Briones. All rights reserved. */
+/* SPDX-License-Identifier: AGPL-3.0-or-later */
+/* This file is part of https://github.com/texsydo/texsydo---mvp */
