+++ tsd.math.software---mvp/src/app/App.test.tsx
@@ -1,5 +1,9 @@
+// Copyright (c) 2024 Tobias Briones. All rights reserved.
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// This file is part of https://github.com/texsydo/texsydo---mvp
+
+import App from "@app/App.tsx";
 import { render, screen } from "@testing-library/react";
-import App from "./App";
 
 describe("App tests", () => {
     it("should render the title", () => {
@@ -9,6 +13,6 @@ describe("App tests", () => {
             screen.getByRole("heading", {
                 level: 1,
             }),
-        ).toHaveTextContent("Vite + React");
+        ).toHaveTextContent("Texsydo");
     });
 });
