+++ tsd.math.software---mvp/src/info.tsx
@@ -1,0 +1,8 @@
+// Copyright (c) 2024 Tobias Briones. All rights reserved.
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// This file is part of https://github.com/texsydo/texsydo---mvp
+
+export const APP_SEMVER = "0.1.0-dev";
+export const APP_NAME = "Texsydo (Tsd.Math.Software)";
+export const APP_VERSION = `MVP v${ APP_SEMVER }`;
+export const APP_VERSION_LABEL = `${ APP_NAME } ${ APP_VERSION }`;
