+++ tsd.math.software---mvp/src/setupTests.ts
@@ -1,3 +1,7 @@
+// Copyright (c) 2024 Tobias Briones. All rights reserved.
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// This file is part of https://github.com/texsydo/texsydo---mvp
+
 import * as matchers from "@testing-library/jest-dom/matchers";
 import { expect } from "vitest";
