+++ tsd.math.software---mvp/vitest.config.ts
@@ -1,3 +1,7 @@
+// Copyright (c) 2024 Tobias Briones. All rights reserved.
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// This file is part of https://github.com/texsydo/texsydo---mvp
+
 import { defineConfig } from "vitest/config";
 import viteConfig from "./vite.config.ts";
